backend/requirements.txt
Django==3.2
ruff==0.0.290
djoser==2.1.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-extra-fields==3.4.0
webcolors==1.12
python-dotenv==0.20.0
reportlab==3.6.11
Pillow==9.0.0
psycopg2-binary==2.9.3
